document-processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0
aiofiles==24.1.0
sqlalchemy==1.4.49
databases[aiosqlite]==0.8.0
python-magic==0.4.27
uuid==1.30
asyncio==3.4.3
typing-extensions==4.14.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas>=2.0.0
langchain==0.1.0
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.30.2
supabase>=2.9.0
postgrest>=0.16.8
python-dotenv==1.0.0
websockets==15.0.1
huggingface-hub==0.15.1
httpx>=0.26,<0.28
openpyxl==3.1.2
xlrd==2.0.1
python-pptx==0.6.23
bcrypt>=4.0.1
PyJWT>=2.8.0
